+++ e2e/bash/test_bash_shell_integration.py
@@ -21,7 +21,7 @@ _complgen_jit_$stem () {{
     local words cword
     _get_comp_words_by_ref -n \"\$COMP_WORDBREAKS\" words cword
     local prefix=\"\${{words[\$cword]}}\"
-    COMPREPLY+=(\$({complgen_binary_path} jit \"{usage_files_dir}/$stem.usage\" bash --comp-wordbreaks=\"\$COMP_WORDBREAKS\" --prefix="\$prefix" -- \"\${{words[@]:1:\$cword-1}}\"))
+    mapfile -t COMPREPLY < <(complgen jit \"{usage_files_dir}/${{stem}}.usage\" bash --comp-wordbreaks=\"\$COMP_WORDBREAKS\" --prefix=\"\$prefix\" -- \"\${{words[@]:1:\$cword-1}}\")
     return 0
 }}
 "
@@ -52,7 +52,7 @@ mycargo +<toolchain>;
 '''
     with temp_usage_file_path(complgen_binary_path, GRAMMAR, 'mycargo') as usage_file_path:
         input = r'''COMP_WORDS=(mycargo +); COMP_CWORD=1; _complgen_jit_mycargo; printf '%s\n' "${COMPREPLY[@]}"'''
-        assert get_sorted_bash_completions(usage_file_path, input) == sorted(['+foo', '+bar'])
+        assert get_sorted_bash_completions(usage_file_path, input) == sorted(['+foo ', '+bar '])
 
 
 def test_comp_wordbreaks_chars(complgen_binary_path: Path):
@@ -62,4 +62,4 @@ mygrep --color "use markers to highlight the matching strings"=<WHEN>;
 '''
     with temp_usage_file_path(complgen_binary_path, GRAMMAR, 'mygrep') as usage_file_path:
         input = r'''COMP_WORDS=(mygrep --color=); COMP_CWORD=1; _complgen_jit_mygrep; printf '%s\n' "${COMPREPLY[@]}"'''
-        assert get_sorted_bash_completions(usage_file_path, input) == sorted(['always', 'never', 'auto'])
+        assert get_sorted_bash_completions(usage_file_path, input) == sorted(['always ', 'never ', 'auto '])